EE-{816/817} extract to/from bytes methods into separate file, tidy up index.ts
